fix(todos): log a capture's derived failure, drop a dead prop, say why a Shaper waits (PLA-228)
Three verifier findings, nothing else.

BLOCKER — acceptance (d) was only half met. A capture's failure is DERIVED
rather than thrown, so no throw site could ever log it: killing a Shaper
mid-pipeline turned the strip red with the real reason while the gateway log
stayed silent, and the reason died with the browser tab. refreshTodoCapture now
warns from inside its existing signature-change guard, which is already the
"this answer is new" test — so one failure writes one line however long a strip
is left polling.

The onTodoCreated prop had no caller anywhere: invalidating the work-items
query is the whole notification, since the board re-reads through the same
query every other writer refreshes it with.

A rate-limited Shaper (status waiting, lastError set) is parked, not broken —
the gateway resumes it unasked — so the stage stays honest and a waitingReason
rides beside it. The strip renders it as a quiet note under the spinner, not
the red slot: an unexplained spinner on a session asleep for an hour reads as a
hang, and wait-or-recapture is the operator's decision to make.

Splitting three files that the additions pushed past the 300-line ratchet:
todo-capture-facts.ts takes the registry-reading half away from the routes,
todo-capture-get.test.ts takes the GET's reporting (and both halves of (d))
away from the POST's refusals, and capture-waiting.test.tsx takes the parked
capture away from the text path.
