+++ packages/jinn/src/gateway/todo-capture-api.ts
@@ -1,11 +1,10 @@
 import crypto from "node:crypto";
 import type { IncomingMessage as HttpRequest, ServerResponse } from "node:http";
-import { createSession, getMessages, getSession, insertMessage, listSessionsByWorkItem, updateSession } from "../sessions/registry.js";
+import { createSession, getSession, insertMessage, updateSession } from "../sessions/registry.js";
 import { enqueueQueueItem } from "../sessions/queue-item-registry.js";
-import { getWorkItem, linkSession, listWorkItems } from "../work-items/store.js";
+import { getWorkItem, linkSession } from "../work-items/store.js";
 import { logger } from "../shared/logger.js";
 import { readJsonBody } from "./http-helpers.js";
-import { sourceSessionId } from "./approval-authority.js";
 import { resolveCallerIdentity } from "./session-comm-guards.js";
 import { UNIDENTIFIED_TOOL_CALL_ERROR, verifySessionCapability } from "../mcp/identity.js";
 import { isTodoId } from "../work-items/id.js";
@@ -13,8 +12,9 @@ import { orgRegistry } from "./org-registry.js";
 import { json, matchRoute, serverError, type ParsedRoute } from "./route-helpers.js";
 import { resolveMessageAudiences } from "./speech-context.js";
 import { preflightSystemEmployee } from "./system-employee-spawn.js";
-import { TODO_DISPATCHER_NAME, TODO_SHAPER_NAME } from "./system-employees.js";
-import { deriveTodoCaptureState, type TodoCaptureFacts, type TodoCaptureState, type TodoCaptureTodoFact } from "./todo-capture-stage.js";
+import { TODO_SHAPER_NAME } from "./system-employees.js";
+import { refreshTodoCapture } from "./todo-capture-facts.js";
+import { type TodoCaptureState } from "./todo-capture-stage.js";
 import { dispatchWebSessionRun } from "./web-session-dispatch.js";
 import type { Employee, Engine } from "../shared/types.js";
 import type { ApiContext } from "./api.js";
@@ -25,8 +25,8 @@ import type { ApiContext } from "./api.js";
  * The POST is the whole autonomous half of the feature — it spawns the Todo
  * Shaper on the operator's raw text and returns immediately. The GET answers
  * how far that capture has got, and answers it by DERIVING the stage from real
- * state (see todo-capture-stage.ts) rather than reading a stage anyone wrote
- * down. That is what makes a reload recover, and what makes it impossible for
+ * state (todo-capture-facts.ts reads it, todo-capture-stage.ts rules on it)
+ * rather than reading a stage anyone wrote down. That is what makes a reload recover, and what makes it impossible for
  * the strip to show progress the system has not made.
  *
  * The spawn deliberately follows the Todo Dispatcher route's recipe step for
@@ -41,94 +41,6 @@ const CAPTURE_TEXT_MAX = 4_000;
  *  capture, so there is no second identifier to keep in step with it, and a
  *  capture survives a restart exactly as long as its session does. */
 export interface TodoCaptureWire extends TodoCaptureState {}
-
-/** Last stage pushed for a capture, so an unchanged recompute stays quiet.
- *  Bounded by the captures a running gateway has seen; a restart clears it and
- *  the next GET re-pushes, which is the correct behaviour rather than a leak. */
-const lastEmitted = new Map<string, string>();
-
-/**
- * The Todo a capture landed ON, if it landed anywhere.
- *
- * A capture that restated an existing Todo creates nothing, so there is no Todo
- * of its own to read. What it leaves instead is a link from its OWN session to
- * the Todo it restated — one field on the session object the caller already
- * holds. `land_on_work_item` is the only thing that writes it on a shaping
- * session, which is what makes reading it a fact rather than a guess about what
- * the Shaper's comment meant.
- *
- * Todos the capture created itself are excluded: those are answered by the
- * ladder, and a capture linked to its own Todo has not restated anything.
- */
-function landedWorkItem(
-  session: ReturnType<typeof getSession>,
-  created: readonly { id: string }[],
-): { id: string; title: string } | null {
-  const landedId = session?.workItemId;
-  if (!landedId || created.some((item) => item.id === landedId)) return null;
-  const item = getWorkItem(landedId);
-  return item ? { id: item.id, title: item.title } : null;
-}
-
-function factsFor(captureId: string, dispatcherEmployee: string, shaperEmployee: string): TodoCaptureFacts {
-  const session = getSession(captureId);
-  // Narrow in SQL to Todos this employee made from a session, then match the
-  // exact capture on provenance. `createdBy` records the EMPLOYEE, so it cannot
-  // tell two captures apart — `sourceRef` names the session, and is the same
-  // link the dispatch authority walk reads.
-  //
-  // `listWorkItems` orders for the board, not by age. The capture's Todo is the
-  // FIRST one its session made, so age is what this needs.
-  const todos = listWorkItems({ source: "session", createdBy: shaperEmployee })
-    .filter((item) => sourceSessionId(item) === captureId)
-    .sort((a, b) => a.createdAt.localeCompare(b.createdAt) || a.id.localeCompare(b.id));
-
-  return {
-    captureId,
-    landedWorkItem: landedWorkItem(session, todos),
-    session: session
-      ? {
-        id: session.id,
-        status: session.status,
-        // "The engine has said something" is the only honest difference between
-        // a session that has started and one that is working.
-        spoke: getMessages(session.id).some((message) => message.role === "assistant"),
-        attemptOutcome: session.attemptOutcome ?? null,
-        lastError: session.lastError ?? null,
-      }
-      : null,
-    todos: todos.map((item): TodoCaptureTodoFact => ({
-      id: item.id,
-      title: item.title,
-      linked: listSessionsByWorkItem(item.id).map((linked) => ({
-        id: linked.id,
-        employee: linked.employee,
-        workflowId: linked.workflowProvenance?.workflowId ?? null,
-        workflowName: linked.workflowProvenance?.workflowName ?? null,
-        workflowRunId: linked.workflowProvenance?.runId ?? null,
-      })),
-    })),
-    dispatcherEmployee,
-    shaperEmployee,
-  };
-}
-
-/** Derive, and push only when the answer moved. The GET stays the source of
- *  truth; this event only saves the browser from polling for a change that has
- *  not happened yet. */
-export function refreshTodoCapture(context: ApiContext, captureId: string): TodoCaptureState {
-  const state = deriveTodoCaptureState(factsFor(captureId, TODO_DISPATCHER_NAME, TODO_SHAPER_NAME));
-  const signature = `${state.stage}:${state.workItemId ?? ""}:${state.error ?? ""}`;
-  if (lastEmitted.get(captureId) !== signature) {
-    lastEmitted.set(captureId, signature);
-    context.emit("todo-capture:stage", {
-      captureId,
-      stage: state.stage,
-      workItemId: state.workItemId,
-    });
-  }
-  return state;
-}
 
 function capturePrompt(text: string): string {
   return [
