+++ packages/web/src/routes/todos/quick-add/__tests__/capture-bar-voice.test.tsx
@@ -59,6 +59,7 @@ function wire(over: Partial<TodoCaptureWire> = {}): TodoCaptureWire {
     routedTo: null,
     extraWorkItemIds: [],
     error: null,
+    waitingReason: null,
     ...over,
   }
 }
